refactor(model): Move payout curve implementation to deeper module
So that we can use a `payout_curve::inverse` module without having to
expose all the ugly implementation details of our inverse payout
curve.
