+++ model/src/payout_curve/inverse/implementation/README.md
