+++ model/src/payout_curve/inverse/implementation/compat.rs
